requirements.txt
absl-py==2.1.0
albumentations==1.4.3
altgraph==0.17.4
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
attrs==23.2.0
auto-py-to-exe==2.43.3
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==7.1.2
colabcode==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
docopt==0.6.2
Eel==0.16.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.3
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
humanfriendly==10.0
idna==3.6
imageio==2.34.0
ipykernel==6.29.4
ipython==8.12.3
ipython-genutils==0.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
json5==0.9.25
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-server==1.24.0
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab==3.0.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
keyboard==0.13.5
kiwisolver==1.4.5
lazy_loader==0.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
nbclassic==0.3.7
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.4.3
networkx==3.3
notebook==6.4.5
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cuda-runtime-cu12==12.4.127
onnx==1.16.0
onnxruntime-gpu==1.17.1
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
pycocotools==2.0.7
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.17.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.3
PyMsgBox==1.0.9
pyngrok==7.1.6
pyparsing==3.1.2
pyperclip==1.8.2
pypiwin32==223
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.1
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.13.0
seaborn==0.13.2
Send2Trash==1.8.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorrt==10.0.0b6
tensorrt-cu12==10.0.0b6
tensorrt-cu12_bindings==10.0.0b6
tensorrt-cu12_libs==10.0.0b6
tensorrt_dispatch==10.0.0b6
tensorrt_dispatch-cu12==10.0.0b6
tensorrt_dispatch-cu12_bindings==10.0.0b6
tensorrt_dispatch-cu12_libs==10.0.0b6
tensorrt_lean==10.0.0b6
tensorrt_lean-cu12==10.0.0b6
tensorrt_lean-cu12_bindings==10.0.0b6
tensorrt_lean-cu12_libs==10.0.0b6
terminado==0.18.1
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2024.2.12
tinycss2==1.2.1
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2+cu121
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.1.44
urllib3==2.2.1
uvicorn==0.13.1
virtualenv==20.25.3
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
whichcraft==0.6.1
yarg==0.1.9
zope.event==5.0
zope.interface==6.2
